Move JsonSaveData to separate file
