+++ db/src/save_data.rs
@@ -1,0 +1,2 @@
+pub mod json_save_data;
+pub use json_save_data::*;
